increase copy timeout
